Second Commit: -Changed Naming for certain classes to make more sense -Receipt now has ReceiptOutputStrategy AND SaleOutputStrategy -Added configurations in the Startup class -Basic "House-keeping" ***Implementations still need to be defined further.***

+++ src/cbg/discountstrategy/NoDiscount.java
@@ -6,7 +6,7 @@ package cbg.discountstrategy;
  * @author cgonz
  */
 public class NoDiscount implements DiscountStrategy {
-    private final int NO_DISCOUNT = 0;
+    private final double NO_DISCOUNT = 0;
     @Override
     public final double getDiscount(double unitPrice, int productQty) {
         return NO_DISCOUNT;
